+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 # 6.3.0-alpha.1
 
 * Feat: Automatically create transactions when navigating between screens (#643)
+* Fix: Do not throw when Throwable type is not supported for associating errors to a transaction (#)
 
 # 6.2.2